+++ app/connectors_service/connectors/sources/google_cloud_storage/gcp_clients.py
@@ -1,0 +1,116 @@
+#
+# Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+# or more contributor license agreements. Licensed under the Elastic License 2.0;
+# you may not use this file except in compliance with the Elastic License 2.0.
+#
+"""Google Cloud Storage source module responsible to fetch documents from Google Cloud Storage."""
+
+import os
+
+from aiogoogle import Aiogoogle
+from aiogoogle.auth.creds import ServiceAccountCreds
+from connectors_sdk.logger import logger
+
+from connectors.utils import RetryStrategy, retryable
+
+CLOUD_STORAGE_READ_ONLY_SCOPE = "https://www.googleapis.com/auth/devstorage.read_only"
+CLOUD_STORAGE_BASE_URL = "https://console.cloud.google.com/storage/browser/_details/"
+API_NAME = "storage"
+API_VERSION = "v1"
+
+RETRY_COUNT = 3
+RETRY_INTERVAL = 2
+STORAGE_EMULATOR_HOST = os.environ.get("STORAGE_EMULATOR_HOST")
+RUNNING_FTEST = (
+    "RUNNING_FTEST" in os.environ
+)  # Flag to check if a connector is run for ftest or not.
+
+
+class GoogleCloudStorageClient:
+    """A google client to handle api calls made to Google Cloud Storage."""
+
+    def __init__(self, json_credentials):
+        """Initialize the ServiceAccountCreds class using which api calls will be made.
+
+        Args:
+            json_credentials (dict): Service account credentials json.
+        """
+        self.service_account_credentials = ServiceAccountCreds(
+            scopes=[CLOUD_STORAGE_READ_ONLY_SCOPE],
+            **json_credentials,
+        )
+        self.user_project_id = self.service_account_credentials.project_id
+        self._logger = logger
+
+    def set_logger(self, logger_):
+        self._logger = logger_
+
+    @retryable(
+        retries=RETRY_COUNT,
+        interval=RETRY_INTERVAL,
+        strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
+    )
+    async def api_call(
+        self,
+        resource,
+        method,
+        sub_method=None,
+        full_response=False,
+        **kwargs,
+    ):
+        """Make a GET call for Google Cloud Storage with retry for the failed API calls.
+
+        Args:
+            resource (aiogoogle.resource.Resource): Resource name for which api call will be made.
+            method (aiogoogle.resource.Method): Method available for the resource.
+            sub_method (aiogoogle.resource.Method, optional): Sub-method available for the method. Defaults to None.
+            full_response (bool, optional): Specifies whether the response is paginated or not. Defaults to False.
+
+        Raises:
+            exception: A instance of an exception class.
+
+        Yields:
+            Dictionary: Response returned by the resource method.
+        """
+        while True:
+            try:
+                async with Aiogoogle(
+                    service_account_creds=self.service_account_credentials
+                ) as google_client:
+                    storage_client = await google_client.discover(
+                        api_name=API_NAME, api_version=API_VERSION
+                    )
+                    if RUNNING_FTEST and not sub_method and STORAGE_EMULATOR_HOST:
+                        self._logger.debug(
+                            f"Using the storage emulator at {STORAGE_EMULATOR_HOST}"
+                        )
+                        # Redirecting calls to fake Google Cloud Storage server for e2e test.
+                        storage_client.discovery_document["rootUrl"] = (
+                            STORAGE_EMULATOR_HOST + "/"
+                        )
+                    resource_object = getattr(storage_client, resource)
+                    method_object = getattr(resource_object, method)
+                    self._logger.debug(
+                        f"Calling '{resource}.{method}' with args: '{dict(**kwargs)}'"
+                    )
+                    if full_response:
+                        first_page_with_next_attached = (
+                            await google_client.as_service_account(
+                                method_object(**kwargs),
+                                full_res=True,
+                            )
+                        )
+                        async for page_items in first_page_with_next_attached:
+                            yield page_items
+                    else:
+                        if sub_method:
+                            method_object = getattr(method_object, sub_method)
+                        yield await google_client.as_service_account(
+                            method_object(**kwargs)
+                        )
+                    break
+            except AttributeError as error:
+                self._logger.error(
+                    f"Error occurred while generating the resource/method object for an API call. Error: {error}"
+                )
+                raise
