+++ app/connectors_service/tests/sources/test_dropbox.py
@@ -22,7 +22,7 @@ from connectors_sdk.filtering.validation import Filter, SyncRuleValidationResult
 from connectors_sdk.source import ConfigurableFieldValueError
 from freezegun import freeze_time
 
-from connectors.sources.dropbox import (
+from app.connectors_service.connectors.sources.dropbox.dropbox_client import (
     AUTHENTICATED_ADMIN_URL,
     DropBoxAdvancedRulesValidator,
     DropboxClient,
