+++ app/connectors_service/tests/sources/test_gmail.py
@@ -15,13 +15,13 @@ from connectors_sdk.source import ConfigurableFieldValueError
 from connectors_sdk.utils import Features, iso_utc
 from freezegun import freeze_time
 
-from connectors.sources.gmail import (
+from app.connectors_service.connectors.sources.gmail.validator import (
     ACCESS_CONTROL,
     GMailAdvancedRulesValidator,
     GMailDataSource,
     _message_doc,
 )
-from connectors.sources.google import MessageFields, UserFields
+from app.connectors_service.connectors.sources.shared.google.google import MessageFields, UserFields
 from tests.commons import AsyncIterator
 from tests.sources.support import create_source
